requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.4
django-filter==22.1
idna==3.4
Pillow==9.3.0
python-dotenv==0.21.0
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
